feat(dashboard): feature-flag switch between legacy and new dashboard
